Add yaml parsing of post meta data
Using yaml as posts meta data markup so it's easier to edit. It's not
fixed to the line order as it was before but to the value in yaml part
of the post. It's on the top of the post and it's separated from the
post body with three dashes.